chg: [views] Check the existance of the collection to remove.

+++ mosp/views/collection.py
@@ -164,10 +164,11 @@ def delete(collection_id=None):
     Delete the requested collection.
     """
     elem = Collection.query.filter(Collection.id == collection_id).first()
-    db.session.delete(elem)
-    db.session.commit()
-    flash(
-        gettext("%(object_name)s successfully deleted.", object_name=elem.name),
-        "success",
-    )
+    if elem:
+        db.session.delete(elem)
+        db.session.commit()
+        flash(
+            gettext("%(object_name)s successfully deleted.", object_name=elem.name),
+            "success",
+        )
     return redirect(url_for("collections_bp.list_collections"))
